+++ core/CHANGELOG.md
@@ -10,6 +10,7 @@
 * Pass network ID instead of network name to the controller
 * Switching FreeIpAddrTracker to pass net.IP objects instead of strings
 * Renaming many parameters and variables to represent network IDs instead of names
+* Change networks.ServiceID to strings instead of int
 
 # 0.7.0
 * Allow developers to configure how wide their test networks will be